:construction: Ajustando erro do unlink no windows com arquivo temporário

+++ funcoes_gerenciador_de_chamados.php
@@ -43,7 +43,10 @@ function salvarChamados($caminhoDoArquivo, $dados){
         }
     }
 
-    unlink($arquivoTemporario);
+    if(file_exists($arquivoTemporario)){
+        unlink($arquivoTemporario);
+
+    }
     return true;
 }
